Restructure codebase (#80)
* Restructure codebase

* Add back import to register environments

* Merge util/trainer.py with algorithms/adversarial.py

* Update script references in experiments/
