+++ .codecov.yml
@@ -6,12 +6,12 @@ coverage:
         target: 80%
         paths:
           - "src/imitation/"
-          - "!src/imitation/examples/"
+          - "!src/imitation/envs/examples/"
           - "!src/imitation/scripts/"
       auxiliary:
         target: 0%
         paths:
-          - "src/imitation/examples/"
+          - "src/imitation/envs/examples/"
           - "src/imitation/scripts/"
       tests:
         # Should not have dead code in our tests
@@ -23,11 +23,11 @@ coverage:
       main:
         paths:
           - "src/imitation/"
-          - "!src/imitation/examples/"
+          - "!src/imitation/envs/examples/"
           - "!src/imitation/scripts/"
       auxiliary:
         paths:
-          - "src/imitation/examples/"
+          - "src/imitation/envs/examples/"
           - "src/imitation/scripts/"
       tests:
         target: 100%
