+++ docs/imitation.algorithms.rst
@@ -1,0 +1,46 @@
+imitation.algorithms package
+============================
+
+Submodules
+----------
+
+imitation.algorithms.adversarial module
+---------------------------------------
+
+.. automodule:: imitation.algorithms.adversarial
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.algorithms.bc module
+------------------------------
+
+.. automodule:: imitation.algorithms.bc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.algorithms.density\_baselines module
+----------------------------------------------
+
+.. automodule:: imitation.algorithms.density_baselines
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.algorithms.tabular\_irl module
+----------------------------------------
+
+.. automodule:: imitation.algorithms.tabular_irl
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.algorithms
+   :members:
+   :undoc-members:
+   :show-inheritance:
