+++ docs/imitation.envs.examples.rst
@@ -1,0 +1,22 @@
+imitation.envs.examples package
+===============================
+
+Submodules
+----------
+
+imitation.envs.examples.model\_envs module
+------------------------------------------
+
+.. automodule:: imitation.envs.examples.model_envs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.envs.examples
+   :members:
+   :undoc-members:
+   :show-inheritance:
