+++ docs/imitation.envs.rst
@@ -1,0 +1,29 @@
+imitation.envs package
+======================
+
+Subpackages
+-----------
+
+.. toctree::
+
+   imitation.envs.examples
+
+Submodules
+----------
+
+imitation.envs.resettable\_env module
+-------------------------------------
+
+.. automodule:: imitation.envs.resettable_env
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.envs
+   :members:
+   :undoc-members:
+   :show-inheritance:
