--- docs/imitation.examples.rst
@@ -1,22 +1,0 @@
-imitation.examples package
-==========================
-
-Submodules
-----------
-
-imitation.examples.model\_envs module
--------------------------------------
-
-.. automodule:: imitation.examples.model_envs
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: imitation.examples
-   :members:
-   :undoc-members:
-   :show-inheritance:
