+++ docs/imitation.policies.rst
@@ -1,0 +1,30 @@
+imitation.policies package
+==========================
+
+Submodules
+----------
+
+imitation.policies.base module
+------------------------------
+
+.. automodule:: imitation.policies.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.policies.serialize module
+-----------------------------------
+
+.. automodule:: imitation.policies.serialize
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.policies
+   :members:
+   :undoc-members:
+   :show-inheritance:
