+++ docs/imitation.rewards.rst
@@ -1,0 +1,30 @@
+imitation.rewards package
+=========================
+
+Submodules
+----------
+
+imitation.rewards.discrim\_net module
+-------------------------------------
+
+.. automodule:: imitation.rewards.discrim_net
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.rewards.reward\_net module
+------------------------------------
+
+.. automodule:: imitation.rewards.reward_net
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.rewards
+   :members:
+   :undoc-members:
+   :show-inheritance:
