+++ docs/imitation.rst
@@ -6,57 +6,20 @@ Subpackages
 
 .. toctree::
 
-   imitation.examples
+   imitation.algorithms
+   imitation.envs
+   imitation.policies
+   imitation.rewards
    imitation.scripts
    imitation.util
 
 Submodules
 ----------
 
-imitation.discrim\_net module
------------------------------
-
-.. automodule:: imitation.discrim_net
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-imitation.resettable\_env module
---------------------------------
-
-.. automodule:: imitation.resettable_env
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-imitation.reward\_net module
-----------------------------
-
-.. automodule:: imitation.reward_net
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 imitation.summaries module
 --------------------------
 
 .. automodule:: imitation.summaries
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-imitation.tabular\_irl module
------------------------------
-
-.. automodule:: imitation.tabular_irl
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-imitation.trainer module
-------------------------
-
-.. automodule:: imitation.trainer
    :members:
    :undoc-members:
    :show-inheritance:
