+++ docs/imitation.scripts.config.rst
@@ -1,0 +1,46 @@
+imitation.scripts.config package
+================================
+
+Submodules
+----------
+
+imitation.scripts.config.common module
+--------------------------------------
+
+.. automodule:: imitation.scripts.config.common
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.scripts.config.eval\_policy module
+--------------------------------------------
+
+.. automodule:: imitation.scripts.config.eval_policy
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.scripts.config.expert\_demos module
+---------------------------------------------
+
+.. automodule:: imitation.scripts.config.expert_demos
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+imitation.scripts.config.train\_adversarial module
+--------------------------------------------------
+
+.. automodule:: imitation.scripts.config.train_adversarial
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: imitation.scripts.config
+   :members:
+   :undoc-members:
+   :show-inheritance:
