+++ experiments/mujoco_experts.sh
@@ -13,5 +13,5 @@ TIMESTAMP=$(${DATE_CMD} --iso-8601=seconds)
 OUTPUT_DIR=output/mujoco_experts/${TIMESTAMP}/
 
 parallel -j 25% --header : --results ${OUTPUT_DIR}/parallel/ \
-         python -m imitation.scripts.data_collect with log_root=${OUTPUT_DIR} {env} seed={seed} \
+         python -m imitation.scripts.expert_demos with log_root=${OUTPUT_DIR} {env} seed={seed} \
          ::: env ${ENVS} ::: seed ${SEEDS}
