+++ src/imitation/algorithms/tabular_irl.py
