+++ src/imitation/envs/examples/airl_envs/assets/twod_maze.xml
