+++ src/imitation/scripts/config/eval_policy.py
@@ -4,10 +4,10 @@ import sacred
 
 from imitation.util import util
 
-policy_eval_ex = sacred.Experiment("policy_eval")
+eval_policy_ex = sacred.Experiment("eval_policy")
 
 
-@policy_eval_ex.config
+@eval_policy_ex.config
 def replay_defaults():
   env_name = "CartPole-v1"  # environment to evaluate in
   timesteps = int(1e4)  # number of timesteps to evaluate
@@ -16,15 +16,15 @@ def replay_defaults():
   render = True  # render to screen
   policy_type = "ppo2"  # class to load policy, see imitation.policies.loader
   policy_path = "expert_models/PPO2_CartPole-v1_0"  # serialized policy
-  log_root = os.path.join("output", "policy_eval")  # output directory
+  log_root = os.path.join("output", "eval_policy")  # output directory
 
 
-@policy_eval_ex.config
+@eval_policy_ex.config
 def logging(log_root, env_name):
   log_dir = os.path.join(log_root, env_name.replace("/", "_"),
                          util.make_unique_timestamp())
 
 
-@policy_eval_ex.named_config
+@eval_policy_ex.named_config
 def fast():
   timesteps = int(1e2)
