+++ src/imitation/scripts/config/expert_demos.py
@@ -5,11 +5,11 @@ import sacred
 from imitation.scripts.config.common import DEFAULT_BLANK_POLICY_KWARGS
 from imitation.util import util
 
-data_collect_ex = sacred.Experiment("data_collect")
+expert_demos_ex = sacred.Experiment("expert_demos")
 
 
-@data_collect_ex.config
-def data_collect_defaults():
+@expert_demos_ex.config
+def expert_demos_defaults():
   env_name = "CartPole-v1"  # The gym.Env name
   total_timesteps = int(1e6)  # Number of training timesteps in model.learn()
   num_vec = 8  # Number of environments in VecEnv
@@ -25,16 +25,16 @@ def data_collect_defaults():
   policy_save_interval = 100  # Num updates between saves (<=0 disables)
   policy_save_final = True  # If True, save after training is finished.
 
-  log_root = os.path.join("output", "data_collect")  # output directory
+  log_root = os.path.join("output", "expert_demos")  # output directory
 
 
-@data_collect_ex.config
+@expert_demos_ex.config
 def logging(env_name, log_root):
   log_dir = os.path.join(log_root, env_name.replace('/', '_'),
                          util.make_unique_timestamp())
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def ant():
   env_name = "Ant-v2"
   make_blank_policy_kwargs = dict(
@@ -43,24 +43,24 @@ def ant():
   total_timesteps = int(5e6)  # OK after 2e6, but continues improving
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def acrobot():
   env_name = "Acrobot-v1"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def cartpole():
   env_name = "CartPole-v1"
   total_timesteps = int(1e5)
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def half_cheetah():
   env_name = "HalfCheetah-v2"
   total_timesteps = int(5e6)  # does OK after 1e6, but continues improving
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def humanoid():
   env_name = "Humanoid-v2"
   make_blank_policy_kwargs = dict(
@@ -69,38 +69,38 @@ def humanoid():
   total_timesteps = int(10e6)  # fairly discontinuous, needs at least 5e6
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def hopper():
   # TODO(adam): upgrade to Hopper-v3?
   env_name = "Hopper-v2"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def mountain_car():
   env_name = "MountainCar-v0"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def pendulum():
   env_name = "Pendulum-v0"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def reacher():
   env_name = "Reacher-v2"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def swimmer():
   env_name = "Swimmer-v2"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def walker():
   env_name = "Walker2d-v2"
 
 
-@data_collect_ex.named_config
+@expert_demos_ex.named_config
 def fast():
   """Intended for testing purposes: small # of updates, ends quickly."""
   total_timesteps = int(1e4)
